- 64 bit fixes
git-svn-id: https://svn.openxp.de/openxp/branches/branch_4_10@7171 65356fb5-50e6-0310-a06b-fbc0adf8d39c

+++ netcall/ncpop3.pas
@@ -355,7 +355,7 @@ begin
   if (not FSupportUIDLs)or(not OnlyNew) then
     result:=Nr
   else
-    result:=Integer(FAvailableUIDLs.Objects[Nr - 1]);
+    result:=Int64(FAvailableUIDLs.Objects[Nr - 1]);
 end;
 
 function TPOP3.Retr(ID: Integer; List: TStringList): boolean;
@@ -396,7 +396,7 @@ begin
   if FSupportUIDLs then
     // mark UIDL as retrieved
     for Nr := 0 to FAvailableUIDLs.Count - 1 do
-      if Integer(FAvailableUIDLs.Objects[Nr]) = ID then
+      if Int64(FAvailableUIDLs.Objects[Nr]) = ID then
       begin
         if UIDLs.IndexOf(FAvailableUIDLs[Nr]) = -1 then
           UIDLs.Add(FAvailableUIDLs[Nr]);
